requirements.txt
alabaster==0.7.12
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptg==0.1.0.7
cryptography==2.6.1
docutils==0.14
future==0.17.1
fuzzywuzzy==0.17.0
hypothesis==4.16.0
idna==2.8
imagesize==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
packaging==19.0
pluggy==0.9.0
py==1.8.0
pyaes==1.6.1
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
pyparsing==2.4.0
pytest==4.4.0
python-Levenshtein==0.12.0
python-telegram-bot==11.1.0
pytz==2019.1
requests==2.21.0
rsa==4.0
six==1.12.0
snowballstemmer==1.2.1
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.0
Telethon==1.6.2
urllib3==1.24.1
